+++ src/hkimjp/datascript.clj
@@ -5,8 +5,8 @@
    [datascript.storage.sql.core :as storage-sql]
    [fast-edn.core :refer [read-string]]
    [time-literals.read-write]
-   [cognitect.transit :as t])
-  (:import [java.io ByteArrayInputStream ByteArrayOutputStream]))
+   #_[cognitect.transit :as t])
+  #_(:import [java.io ByteArrayInputStream ByteArrayOutputStream]))
 
 (def datasource
   (doto (org.sqlite.SQLiteDataSource.)
@@ -44,35 +44,3 @@
 (defn close-conn []
   (storage-sql/close storage)
   (alter-var-root #'conn (constantly nil)))
-
-;------
-
-; (create-conn schema storage)
-
-; ; (def conn (d/create-conn schema {:storage storage}))
-
-; (d/transact! conn [{:db/id -1, :time (java.util.Date.)}])
-; (d/transact! conn [{:db/id -1, :time (jt/local-date-time)}])
-
-; (storage-sql/close storage)
-
-; (restore-conn storage)
-
-; (d/transact! conn [{:db/id -1, :name "hiroshi"}])
-; (d/transact! conn [{:db/id -1, :date (jt/local-date)}])
-; (d/transact! conn [{:db/id -1, :name "kimura" :date (jt/local-date)}])
-
-; (d/q '[:find ?e ?time
-;        :where
-;        [?e :time ?time]]
-;      @conn)
-
-; (d/q '[:find ?e ?time
-;        :where
-;        [?e :date ?time]]
-;      @conn)
-
-; (d/pull @conn '[*] 2)
-
-; (close)
-